moved accordion build to it's own module
